Remove useless conversions
Fixes clippy::useless_conversion

+++ src/drivers/virtio/virtqueue/split.rs
@@ -334,7 +334,7 @@ impl SplitVq {
 			let mut index = index.iter();
 			// Even on 64bit systems this is fine, as we have a queue_size < 2^15!
 			let det_notif_data: u16 = (next_off as u16) >> 1;
-			let flags = (det_notif_data | (u16::from(next_wrap) << 15)).to_le_bytes();
+			let flags = (det_notif_data | (next_wrap << 15)).to_le_bytes();
 			let mut flags = flags.iter();
 			let mut notif_data: [u8; 4] = [0, 0, 0, 0];
